dagster_ai4ef_train_app/python_requirements.txt
click==8.1.7
fastapi==0.104.1
matplotlib==3.8.2
numpy==1.26.2
optuna==3.4.0
pandas==2.1.3
protobuf==4.25.1
python-dotenv==1.0.0
pytorch-lightning==2.1.2
pytz==2023.3.post1
scikit-learn==1.3.2
seaborn==0.13.0
statsmodels==0.14.0
torchensemble==0.1.9
torchmetrics==1.2.1
uvicorn==0.24.0.post1
u8darts[torch]
plotly # for optuna visualization plots
tornado # for webagg on optuna viz (not necessary for non-GUI env)
SQLAlchemy==2.0.23
httpx==0.25.2
psycopg2-binary==2.9.9
dagster==1.6.3 
dagster-cloud==1.6.3 
dagster-webserver==1.6.3